+++ Changelog.md
@@ -3,6 +3,7 @@ Lilu Changelog
 #### v1.5.8
 - Fixed kernel panic on macOS 10.15 and earlier introduced in 1.5.7
 - Added Alder Lake CPU model support
+- Added shared patcher instance grabbing API
 
 #### v1.5.7
 - Added address slot support for all 64-bit macOS version