+++ Changelog.md
@@ -4,6 +4,7 @@ Lilu Changelog
 - Fixed kernel panic on macOS 10.15 and earlier introduced in 1.5.7
 - Added Alder Lake CPU model support
 - Added shared patcher instance grabbing API
+- Fixed memory corruption when mixing cs_validate_range/page mid/long routes (thx @Goshin)
 
 #### v1.5.7
 - Added address slot support for all 64-bit macOS version